pyproject.toml
[project]
name = "lp_problem"
version = "0.1.0"
description = ""
authors = [
  {name = "Mihhail Samusev", email = "" }
]
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    "Pyomo==6.6",
    "pydantic==2.*",
    "python-decouple",
]


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
